adding files, the parseToken function is bringing the 403 auth header
